This code snippet sets the background image of an element using a linear gradient and an image.
The linear gradient is specified using the linear-gradient() function and has an angle of 105 degrees. It consists of two color stops - the first at 0% with an opacity of 0.9 for a white color (rgba($color-white, .9)) and the second at 65% with the same color and opacity. This creates a white-to-transparent gradient that covers the top portion of the background image.

The url() function is then used to specify the image file (../img/nat-10.jpg) that will be used as the background image. This image will be placed below the gradient and will fill the remaining portion of the element's background.

Overall, this code snippet creates a layered background effect with a gradient on top of an image.