This code snippet sets the background image of an element using a linear gradient and an image for a specific screen size range.
The @include rule is a Sass/SCSS directive that includes a mixin or a function in the current context. In this case, it includes a mixin named respond with a parameter of tab-land. This suggests that this mixin is used to apply styles for a specific screen size range, which is commonly referred to as "tablet landscape" size.

Within the respond(tab-land) block, the code sets the background image of the element with a linear gradient and an image. The linear gradient is specified using the linear-gradient() function and has a direction of "to right", which creates a horizontal gradient. It consists of two color stops - the first at 0% with an opacity of 0.9 for a white color (rgba($color-white, .9)) and the second at 100% with the same color and opacity. This creates a solid white color that covers the entire background area.

The url() function is then used to specify the image file (../img/nat-10.jpg) that will be used as the background image. This image will be placed on top of the white gradient and will fill the entire background area.

Overall, this code snippet creates a white-to-image gradient effect for the element for tablet landscape screen size range.